requirements.txt
flask
qrcode
pillow
fpdf
gunicorn